db_api/requirements.txt
python-dotenv==0.20.0
asgiref==3.5.2
django
sqlparse==0.4.2
djangorestframework==3.13.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
